# .history/docs/design/cash-flow-api-specification_20250905215824.yaml
openapi: 3.0.3
info:
  title: Cash Flow Management Service API
  description: |
    Pure calculation service API for cash flows in synthetic swap contracts.
    Focuses on calculation performance with zero storage overhead for input data.
    Supports real-time processing, incremental updates, and historical recalculation.
  version: 1.0.0
  contact:
    name: Core Lifecycle Management Team
    email: lifecycle-team@financial-firm.com
  license:
    name: Proprietary
    url: https://financial-firm.com/license
servers:
  - url: https://cashflow-service.internal/api/v1
    description: Internal production server
  - url: https://cashflow-service-staging.internal/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /cashflows/calculate:
    post:
      tags:
        - Calculations
      summary: Calculate cash flows (auto-determined type)
      description: |
        Main endpoint for cash flow calculations. The service automatically determines the calculation type based on:
        - Date range size (historical vs real-time)
        - Market data availability (incremental vs full)
        - Request context and performance requirements
        
        Calculation Type Logic:
        - REAL_TIME_PROCESSING: Single day or current date calculations (< 24 hours)
        - INCREMENTAL_UPDATE: Small date ranges (1-7 days) with existing market data
        - HISTORICAL_RECALCULATION: Large date ranges (> 7 days) or full recalculation
        
        Pure calculation service with zero storage overhead.
      operationId: calculateCashFlows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
            examples:
              realTimeExample:
                summary: Real-time calculation example (auto-determined)
                value:
                  cashFlowRequest:
                    requestId: "CF_REQ_RT_001"
                    contractId: "SWAP_IBM_001"
                    dateRange:
                      fromDate: "2024-01-15"
                      toDate: "2024-01-15"
                      calculationFrequency: "DAILY"
                    contracts:
                      - contractId: "SWAP_IBM_001"
                        underlying: "IBM"
                        notionalAmount: 400000
                        currency: "USD"
                        equityLeg:
                          underlying: "IBM"
                          quantity: 1000
                          dividendTreatment: "REINVEST"
                        interestLeg:
                          rateType: "FLOATING"
                          index: "LIBOR_3M"
                          spread: 0.0025
                          resetFrequency: "QUARTERLY"
                    positions:
                      - positionId: "POS_IBM_001"
                        contractId: "SWAP_IBM_001"
                        totalQuantity: 1000
                        status: "ACTIVE"
                    lots:
                      - lotId: "LOT_001"
                        contractId: "SWAP_IBM_001"
                        quantity: 1000
                        costPrice: 120.00
                        costDate: "2019-01-01"
                        status: "ACTIVE"
                    paymentSchedules:
                      - scheduleId: "SCH_001"
                        contractId: "SWAP_IBM_001"
                        scheduleType: "INTEREST_PAYMENT"
                        scheduledDate: "2024-01-15"
                        leg: "INTEREST"
                        notionalAmount: 400000
                    marketDataStrategy:
                      mode: "HYBRID"
                      realTimeEndpoints:
                        enabled: true
                        timeoutSeconds: 10
                      cacheStrategy:
                        enabled: true
                        ttlHours: 24
                    marketData:
                      mode: "SELF_CONTAINED"
                      data:
                        securities:
                          - symbol: "IBM"
                            basePrice: 125.50
                            baseDate: "2024-01-15"
                            changes:
                              - date: "2024-01-15"
                                price: 125.50
                                change: 0.75
                        rates:
                          - index: "LIBOR_3M"
                            baseRate: 0.0525
                            baseDate: "2024-01-15"
                            changes:
                              - date: "2024-01-15"
                                rate: 0.0525
                                change: 0.0010
                        dividends:
                          - symbol: "IBM"
                            dividends:
                              - exDate: "2024-03-15"
                                paymentDate: "2024-03-30"
                                amount: 1.65
                                currency: "USD"
      responses:
        '200':
          description: Cash flow calculation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowResponse'
              examples:
                successExample:
                  summary: Successful calculation response
                  value:
                    requestId: "CF_REQ_RT_001"
                    calculationDate: "2024-01-15"
                    calculationType: "REAL_TIME_PROCESSING"
                    status: "COMPLETED"
                    summary:
                      totalContracts: 1
                      totalCashFlows: 2
                      totalAmount: 1250.75
                      currency: "USD"
                    contractResults:
                      - contractId: "SWAP_IBM_001"
                        underlying: "IBM"
                        totalPnl: 550.25
                        totalInterest: 525.50
                        totalDividends: 175.00
                        totalCashFlows: 1250.75
                        cashFlows:
                          - cashFlowId: "CF_001"
                            calculationDate: "2024-01-15"
                            cashFlowType: "INTEREST_ACCRUAL"
                            equityLegAmount: 0.00
                            interestLegAmount: 525.50
                            totalAmount: 525.50
                            currency: "USD"
                            state: "REALIZED_UNSETTLED"
                          - cashFlowId: "CF_002"
                            calculationDate: "2024-01-15"
                            cashFlowType: "EQUITY_PNL"
                            equityLegAmount: 550.25
                            interestLegAmount: 0.00
                            totalAmount: 550.25
                            currency: "USD"
                            state: "REALIZED_UNSETTLED"
                    metadata:
                      calculationVersion: "1.0"
                      calculationEngine: "PureCalculationEngine"
                      processingTimeMs: 280
                      memoryUsageMB: 512
                      dataSource: "HYBRID"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CashFlowRequest:
      type: object
      required:
        - requestId
        - dateRange
        - contracts
        - marketDataStrategy
        - marketData
      properties:
        requestId:
          type: string
          description: Unique identifier for the calculation request (synthetic key)
          example: "CF_REQ_001"
        underlying:
          type: string
          description: Top-level underlying asset symbol for market data extraction
          example: "IBM"
        index:
          type: string
          description: Top-level interest rate index for market data extraction
          example: "LIBOR_3M"
        calculationType:
          type: string
          enum: [REAL_TIME_PROCESSING, BATCH_PROCESSING, HISTORICAL_RECALCULATION]
          description: Type of calculation to perform
          example: "REAL_TIME_PROCESSING"
        dateRange:
          $ref: '#/components/schemas/DateRange'
        contracts:
          type: array
          items:
            $ref: '#/components/schemas/Contract'
          description: Contract details for calculation
        marketDataStrategy:
          $ref: '#/components/schemas/MarketDataStrategy'
        marketData:
          $ref: '#/components/schemas/MarketDataContainer'
...
    DateRange:
      type: object
      required:
        - fromDate
        - toDate
      properties:
        fromDate:
          type: string
          format: date
          description: Start date for calculations
          example: "2024-01-01"
        toDate:
          type: string
          format: date
          description: End date for calculations
          example: "2024-01-15"
        calculationFrequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, QUARTERLY]
          description: Frequency of calculations
          example: "DAILY"
    Contract:
      type: object
      required:
        - contractId
        - underlying
        - index
        - notionalAmount
        - currency
        - type
        - startDate
        - endDate
      properties:
        contractId:
          type: string
          description: Unique contract identifier
          example: "EQ_SWAP_001"
        underlying:
          type: string
          description: Underlying asset symbol
          example: "IBM"
        index:
          type: string
          description: Interest rate index
          example: "LIBOR_3M"
        notionalAmount:
          type: number
          format: double
          description: Contract notional amount
          example: 1000000.0
        currency:
          type: string
          description: Contract currency
          example: "USD"
        type:
          type: string
          enum: [EQUITY_SWAP, INTEREST_RATE_SWAP, BOND, OPTION, FUTURE, FORWARD]
          description: Contract type
          example: "EQUITY_SWAP"
        startDate:
          type: string
          format: date
          description: Contract start date
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          description: Contract end date
          example: "2024-12-31"
...
    MarketDataStrategy:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum: [HYBRID, SELF_CONTAINED, ENDPOINTS]
          description: Market data loading strategy
          example: "HYBRID"
        realTimeEndpoints:
          $ref: '#/components/schemas/RealTimeEndpoints'
        cacheStrategy:
          $ref: '#/components/schemas/CacheStrategy'
    RealTimeEndpoints:
      type: object
      properties:
        enabled:
          type: boolean
          description: Enable real-time endpoint calls
          example: true
        timeoutSeconds:
          type: integer
          description: Timeout for endpoint calls
          example: 10
        retryCount:
          type: integer
          description: Number of retry attempts
          example: 3
        endpoints:
          type: object
          properties:
            priceEndpoint:
              type: string
              description: Price data endpoint
              example: "https://market-data.internal/prices"
            rateEndpoint:
              type: string
              description: Rate data endpoint
              example: "https://market-data.internal/rates"
            dividendEndpoint:
              type: string
              description: Dividend data endpoint
              example: "https://market-data.internal/dividends"
    CacheStrategy:
      type: object
      properties:
        enabled:
          type: boolean
          description: Enable market data caching
          example: true
        ttlHours:
          type: integer
          description: Cache time-to-live in hours
          example: 24
        maxSizeMB:
          type: integer
          description: Maximum cache size in MB
          example: 1000
    MarketDataContainer:
      type: object
      required:
        - data
        - version
        - asOfDate
      properties:
        data:
          $ref: '#/components/schemas/MarketDataContent'
        version:
          type: string
          description: Market data version
          example: "1.0"
        asOfDate:
          type: string
          format: date
          description: Market data as-of date
          example: "2024-01-01"
    MarketDataContent:
      type: object
      properties:
        securities:
          type: array
          items:
            $ref: '#/components/schemas/SecurityData'
          description: Security price data
        rates:
          type: array
          items:
            $ref: '#/components/schemas/RateData'
          description: Interest rate data
        dividends:
          type: array
          items:
            $ref: '#/components/schemas/DividendData'
          description: Dividend data
    SecurityData:
      type: object
      required:
        - symbol
        - basePrice
        - baseDate
        - changes
      properties:
        symbol:
          type: string
          description: Security symbol
          example: "IBM"
        basePrice:
          type: number
          format: double
          description: Base price
          example: 150.0
        baseDate:
          type: string
          format: date
          description: Base date
          example: "2024-01-01"
        changes:
          type: array
          items:
            $ref: '#/components/schemas/PriceChange'
          description: Price changes over time (can be empty array)
          example: []
    RateData:
      type: object
      required:
        - index
        - baseRate
        - baseDate
        - changes
      properties:
        index:
          type: string
          description: Rate index
          example: "LIBOR_3M"
        baseRate:
          type: number
          format: double
          description: Base rate
          example: 0.05
        baseDate:
          type: string
          format: date
          description: Base date
          example: "2024-01-01"
        changes:
          type: array
          items:
            $ref: '#/components/schemas/RateChange'
          description: Rate changes over time (can be empty array)
          example: []
    DividendData:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
          description: Security symbol
          example: "IBM"
        dividends:
          type: array
          items:
            $ref: '#/components/schemas/Dividend'
          description: Dividend information
...
    CashFlowResponse:
      type: object
      required:
        - requestId
        - calculationDate
        - calculationType
        - status
        - summary
        - contractResults
        - metadata
      properties:
        requestId:
          type: string
          description: Original request identifier
          example: "CF_REQ_001"
        calculationDate:
          type: string
          format: date
          description: Calculation date
          example: "2024-01-15"
        calculationType:
          $ref: '#/components/schemas/CalculationTypeDetermination'
          description: Calculation type determination details
        dateRange:
          $ref: '#/components/schemas/DateRange'
        status:
          type: string
          enum: [COMPLETED, PROCESSING, FAILED]
          description: Calculation status
        summary:
          $ref: '#/components/schemas/CalculationSummary'
        contractResults:
          type: array
          items:
            $ref: '#/components/schemas/ContractResult'
          description: Results for each contract
        metadata:
          $ref: '#/components/schemas/CalculationMetadata'
    CalculationSummary:
      type: object
      required:
        - totalContracts
        - totalCashFlows
        - totalAmount
        - currency
      properties:
        totalContracts:
          type: integer
          description: Total number of contracts processed
          example: 1
        totalCashFlows:
          type: integer
          description: Total number of cash flows generated
          example: 2
        totalAmount:
          type: number
          format: double
          description: Total cash flow amount
          example: 1250.75
        currency:
          type: string
          description: Summary currency
          example: "USD"
    ContractResult:
      type: object
      required:
        - contractId
        - underlying
        - totalCashFlows
      properties:
        contractId:
          type: string
          description: Contract identifier
          example: "SWAP_IBM_001"
        underlying:
          type: string
          description: Underlying asset
          example: "IBM"
        totalPnl:
          type: number
          format: double
          description: Total P&L
          example: 550.25
        totalInterest:
          type: number
          format: double
          description: Total interest
          example: 525.50
        totalDividends:
          type: number
          format: double
          description: Total dividends
          example: 175.00
        totalCashFlows:
          type: number
          format: double
          description: Total cash flows
          example: 1250.75
        cashFlows:
          type: array
          items:
            $ref: '#/components/schemas/CashFlow'
          description: Individual cash flows
...
    CalculationMetadata:
      type: object
      required:
        - calculationVersion
        - calculationEngine
        - processingTimeMs
        - dataSource
      properties:
        calculationVersion:
          type: string
          description: Calculation engine version
          example: "1.0"
        calculationEngine:
          type: string
          description: Calculation engine name
          example: "PureCalculationEngine"
        processingTimeMs:
          type: integer
          description: Processing time in milliseconds
          example: 280
        memoryUsageMB:
          type: integer
          description: Memory usage in MB
          example: 512
        dataSource:
          type: string
          enum: [HYBRID, SELF_CONTAINED, ENDPOINTS]
          description: Data source used
          example: "HYBRID"
...
    CalculationTypeDetermination:
      type: object
      required:
        - determinedType
        - reasoning
      properties:
        determinedType:
          type: string
          enum: [REAL_TIME_PROCESSING, INCREMENTAL_UPDATE, HISTORICAL_RECALCULATION]
          description: Automatically determined calculation type
          example: "HISTORICAL_RECALCULATION"
        reasoning:
          type: object
          properties:
            dateRangeDays:
              type: integer
              description: Number of days in the date range
              example: 15
            marketDataMode:
              type: string
              description: Market data loading mode
              example: "SELF_CONTAINED"
            performanceRequirement:
              type: string
              description: Performance requirement that influenced the decision
              example: "LARGE_DATE_RANGE"
            overrideReason:
              type: string
              description: Reason if calculation type was overridden
              example: "EXPLICIT_REQUEST"
        originalRequestedType:
          type: string
          enum: [REAL_TIME_PROCESSING, INCREMENTAL_UPDATE, HISTORICAL_RECALCULATION]
          description: Originally requested calculation type (if provided)
          example: "REAL_TIME_PROCESSING"
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
          description: Error code
          example: "CALCULATION_ERROR"
        message:
          type: string
          description: Error message
          example: "Calculation failed due to invalid market data"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        details:
          type: object
          description: Additional error details
        requestId:
          type: string
          description: Associated request identifier
          example: "CF_REQ_001"