# .history/src/main/java/com/financial/cashflow/controller/CashFlowController_20250903225654.java
package com.financial.cashflow.controller;

import com.financial.cashflow.domain.CashFlowRequestContent;
import com.financial.cashflow.domain.CashFlowResponse;
import com.financial.cashflow.service.CashFlowCalculationService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;
import java.time.LocalDate;
import java.util.List;
import java.util.concurrent.CompletableFuture;

/**
 * REST Controller for Cash Flow Management Service.
 * Provides endpoints for cash flow calculations, batch processing, and status queries.
 * Uses virtual threads for I/O operations and platform threads for CPU work.
 */
@RestController
@RequestMapping("/api/v1/cashflows")
@Tag(name = "Cash Flow Management", description = "Cash flow calculation and management APIs")
@Slf4j
public class CashFlowController {

    @Autowired
    private CashFlowCalculationService calculationService;

    @PostMapping("/calculate")
    @Operation(
        summary = "Calculate cash flows",
        description = "Calculate cash flows for a single contract"
    )
    @ApiResponses(value = {
        @ApiResponse(
            responseCode = "200",
            description = "Calculation completed successfully",
            content = @Content(schema = @Schema(implementation = CashFlowResponse.class))
        ),
        @ApiResponse(
            responseCode = "400",
            description = "Invalid request parameters",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
        ),
        @ApiResponse(
            responseCode = "500",
            description = "Internal server error",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
        )
    })
    public CompletableFuture<ResponseEntity<CashFlowResponse>> calculateCashFlows(
            @Valid @RequestBody CashFlowRequestContent request) {
        
        log.info("Received cash flow calculation request: {}", request.getRequestId());
        
        return calculationService.calculateCashFlows(request)
            .thenApply(response -> {
                log.info("Calculation completed for request: {}", request.getRequestId());
                return ResponseEntity.ok(response);
            })
            .exceptionally(throwable -> {
                log.error("Calculation failed for request: {}", request.getRequestId(), throwable);
                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(CashFlowResponse.builder()
                        .requestId(request.getRequestId())
                        .contractId(request.getContractId())
                        .status("FAILED")
                        .errorMessage(throwable.getMessage())
                        .build());
            });
    }

    @PostMapping("/calculate/batch")
    @Operation(
        summary = "Calculate cash flows in batch",
        description = "Calculate cash flows for multiple contracts in parallel"
    )
    @ApiResponses(value = {
        @ApiResponse(
            responseCode = "200",
            description = "Batch calculation completed successfully",
            content = @Content(schema = @Schema(implementation = BatchResponse.class))
        ),
        @ApiResponse(
            responseCode = "400",
            description = "Invalid request parameters",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
        )
    })
    public CompletableFuture<ResponseEntity<BatchResponse>> calculateCashFlowsBatch(
            @Valid @RequestBody List<CashFlowRequestContent> requests) {
        
        log.info("Received batch cash flow calculation request for {} contracts", requests.size());
        
        List<CompletableFuture<CashFlowResponse>> futures = calculationService.calculateCashFlowsBatch(requests);
        
        return CompletableFuture.allOf(futures.toArray(new CompletableFuture[0]))
            .thenApply(v -> {
                List<CashFlowResponse> responses = futures.stream()
                    .map(CompletableFuture::join)
                    .toList();
                
                log.info("Batch cash flow calculation completed for {} contracts", requests.size());
                return ResponseEntity.ok(BatchResponse.builder()
                    .totalRequests(requests.size())
                    .successfulResponses(responses.size())
                    .responses(responses)
                    .build());
            })
            .exceptionally(throwable -> {
                log.error("Error in batch cash flow calculation", throwable);
                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(BatchResponse.builder()
                        .totalRequests(requests.size())
                        .successfulResponses(0)
                        .errorMessage(throwable.getMessage())
                        .build());
            });
    }

    @PostMapping("/recalculate")
    @Operation(
        summary = "Recalculate cash flows",
        description = "Recalculate cash flows for a specific contract and date range"
    )
    @ApiResponses(value = {
        @ApiResponse(
            responseCode = "200",
            description = "Recalculation completed successfully",
            content = @Content(schema = @Schema(implementation = CashFlowResponse.class))
        ),
        @ApiResponse(
            responseCode = "400",
            description = "Invalid parameters",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
        )
    })
    public CompletableFuture<ResponseEntity<CashFlowResponse>> recalculateCashFlows(
            @Parameter(description = "Contract ID") @RequestParam String contractId,
            @Parameter(description = "From date (yyyy-MM-dd)") @RequestParam String fromDate,
            @Parameter(description = "To date (yyyy-MM-dd)") @RequestParam String toDate) {
        
        log.info("Received recalculation request for contract: {} from {} to {}", 
            contractId, fromDate, toDate);
        
        // Create a request for recalculation
        CashFlowRequestContent request = CashFlowRequestContent.builder()
            .requestId("RECALC_" + contractId + "_" + System.currentTimeMillis())
            .contractId(contractId)
            .dateRange(CashFlowRequestContent.DateRange.builder()
                .fromDate(LocalDate.parse(fromDate))
                .toDate(LocalDate.parse(toDate))
                .calculationFrequency("DAILY")
                .build())
            .calculationType("HISTORICAL_RECALCULATION")
            .build();
        
        return calculationService.recalculateCashFlows(request)
            .thenApply(response -> {
                log.info("Recalculation completed for contract: {}", contractId);
                return ResponseEntity.ok(response);
            })
            .exceptionally(throwable -> {
                log.error("Error recalculating cash flows for contract: {}", contractId, throwable);
                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(CashFlowResponse.builder()
                        .requestId("RECALC_" + contractId)
                        .contractId(contractId)
                        .status("FAILED")
                        .errorMessage(throwable.getMessage())
                        .build());
            });
    }

    @GetMapping("/cached/{requestId}")
    @Operation(
        summary = "Get cached calculation",
        description = "Retrieve a cached cash flow calculation result"
    )
    @ApiResponses(value = {
        @ApiResponse(
            responseCode = "200",
            description = "Cached result found",
            content = @Content(schema = @Schema(implementation = CashFlowResponse.class))
        ),
        @ApiResponse(
            responseCode = "404",
            description = "Cached result not found",
            content = @Content(schema = @Schema(implementation = ErrorResponse.class))
        )
    })
    public ResponseEntity<CashFlowResponse> getCachedCalculation(
            @Parameter(description = "Request ID") @PathVariable String requestId) {
        
        log.info("Retrieving cached calculation for request: {}", requestId);
        
        CashFlowResponse response = calculationService.getCachedCalculation(requestId);
        if (response != null) {
            return ResponseEntity.ok(response);
        } else {
            return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(CashFlowResponse.builder()
                    .requestId(requestId)
                    .status("NOT_FOUND")
                    .errorMessage("Cached calculation not found")
                    .build());
        }
    }

    @GetMapping("/health")
    @Operation(
        summary = "Health check",
        description = "Check the health status of the cash flow service"
    )
    @ApiResponses(value = {
        @ApiResponse(
            responseCode = "200",
            description = "Service is healthy",
            content = @Content(schema = @Schema(implementation = HealthResponse.class))
        )
    })
    public ResponseEntity<HealthResponse> health() {
        log.debug("Health check requested");
        
        return ResponseEntity.ok(HealthResponse.builder()
            .status("UP")
            .timestamp(System.currentTimeMillis())
            .service("cash-flow-management-service")
            .version("1.0.0")
            .build());
    }

    // Response classes for API documentation
    @Schema(description = "Error response")
    public static class ErrorResponse {
        @Schema(description = "Error code")
        private String errorCode;
        
        @Schema(description = "Error message")
        private String message;
        
        @Schema(description = "Timestamp")
        private long timestamp;
        
        // Getters and setters
        public String getErrorCode() { return errorCode; }
        public void setErrorCode(String errorCode) { this.errorCode = errorCode; }
        public String getMessage() { return message; }
        public void setMessage(String message) { this.message = message; }
        public long getTimestamp() { return timestamp; }
        public void setTimestamp(long timestamp) { this.timestamp = timestamp; }
    }

    @Schema(description = "Batch response")
    public static class BatchResponse {
        @Schema(description = "Total number of requests")
        private int totalRequests;
        
        @Schema(description = "Number of successful responses")
        private int successfulResponses;
        
        @Schema(description = "List of responses")
        private List<CashFlowResponse> responses;
        
        @Schema(description = "Error message if any")
        private String errorMessage;
        
        // Builder pattern
        public static BatchResponseBuilder builder() {
            return new BatchResponseBuilder();
        }
        
        public static class BatchResponseBuilder {
            private BatchResponse response = new BatchResponse();
            
            public BatchResponseBuilder totalRequests(int totalRequests) {
                response.totalRequests = totalRequests;
                return this;
            }
            
            public BatchResponseBuilder successfulResponses(int successfulResponses) {
                response.successfulResponses = successfulResponses;
                return this;
            }
            
            public BatchResponseBuilder responses(List<CashFlowResponse> responses) {
                response.responses = responses;
                return this;
            }
            
            public BatchResponseBuilder errorMessage(String errorMessage) {
                response.errorMessage = errorMessage;
                return this;
            }
            
            public BatchResponse build() {
                return response;
            }
        }
        
        // Getters
        public int getTotalRequests() { return totalRequests; }
        public int getSuccessfulResponses() { return successfulResponses; }
        public List<CashFlowResponse> getResponses() { return responses; }
        public String getErrorMessage() { return errorMessage; }
    }

    @Schema(description = "Health response")
    public static class HealthResponse {
        @Schema(description = "Service status")
        private String status;
        
        @Schema(description = "Timestamp")
        private long timestamp;
        
        @Schema(description = "Service name")
        private String service;
        
        @Schema(description = "Service version")
        private String version;
        
        // Builder pattern
        public static HealthResponseBuilder builder() {
            return new HealthResponseBuilder();
        }
        
        public static class HealthResponseBuilder {
            private HealthResponse response = new HealthResponse();
            
            public HealthResponseBuilder status(String status) {
                response.status = status;
                return this;
            }
            
            public HealthResponseBuilder timestamp(long timestamp) {
                response.timestamp = timestamp;
                return this;
            }
            
            public HealthResponseBuilder service(String service) {
                response.service = service;
                return this;
            }
            
            public HealthResponseBuilder version(String version) {
                response.version = version;
                return this;
            }
            
            public HealthResponse build() {
                return response;
            }
        }
        
        // Getters
        public String getStatus() { return status; }
        public long getTimestamp() { return timestamp; }
        public String getService() { return service; }
        public String getVersion() { return version; }
    }
}